+++ python_modules/dagster/dagster_tests/storage_tests/test_run_storage.py
@@ -3,7 +3,7 @@ from contextlib import contextmanager
 
 import mock
 import pytest
-from dagster_tests.core_tests.storage_tests.utils.run_storage import TestRunStorage
+from dagster_tests.storage_tests.utils.run_storage import TestRunStorage
 
 from dagster._core.storage.legacy_storage import LegacyRunStorage
 from dagster._core.storage.runs import InMemoryRunStorage, SqliteRunStorage
