+++ python_modules/libraries/dagster-mysql/dagster_mysql_tests/test_event_log.py
@@ -4,7 +4,7 @@ from urllib.parse import urlparse
 import pytest
 import yaml
 from dagster_mysql.event_log import MySQLEventLogStorage
-from dagster_tests.core_tests.storage_tests.utils.event_log_storage import (
+from dagster_tests.storage_tests.utils.event_log_storage import (
     TestEventLogStorage,
     create_test_event_log_record,
 )
