+++ python_modules/libraries/dagster-mysql/dagster_mysql_tests/test_run_storage.py
@@ -3,7 +3,7 @@ from urllib.parse import urlparse
 import pytest
 import yaml
 from dagster_mysql.run_storage import MySQLRunStorage
-from dagster_tests.core_tests.storage_tests.utils.run_storage import TestRunStorage
+from dagster_tests.storage_tests.utils.run_storage import TestRunStorage
 
 from dagster._core.test_utils import environ, instance_for_test
