+++ python_modules/libraries/dagster-postgres/dagster_postgres_tests/test_run_storage.py
@@ -1,7 +1,7 @@
 import pytest
 import yaml
 from dagster_postgres.run_storage import PostgresRunStorage
-from dagster_tests.core_tests.storage_tests.utils.run_storage import TestRunStorage
+from dagster_tests.storage_tests.utils.run_storage import TestRunStorage
 
 from dagster._core.test_utils import environ, instance_for_test
